HW3/booking_service/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
grpcio==1.62.1
grpcio-tools==1.62.1
SQLAlchemy==2.0.28
alembic==1.13.1
psycopg2-binary==2.9.9
tenacity==8.2.3
email-validator
